requirements.txt
fastapi==0.73.0
uvicorn==0.17.4
pydantic[email]
motor
python-jose[cryptography]
passlib[bcrypt]
python-multipart